chore: refactor more
