+++ src/__tests__/timer-utils.ts
