+++ src/__tests__/timers.test.ts
@@ -1,4 +1,4 @@
-import waitFor from '../waitFor';
+import waitFor from '../wait-for';
 
 describe.each([false, true])(
   'fake timers tests (legacyFakeTimers = %s)',
