+++ src/helpers/__tests__/include-hidden-elements.test.tsx
