+++ src/helpers/string-validation.ts
