+++ src/pure.ts
@@ -1,9 +1,9 @@
 export { default as act } from './act';
 export { default as cleanup } from './cleanup';
-export { default as fireEvent } from './fireEvent';
+export { default as fireEvent } from './fire-event';
 export { default as render } from './render';
-export { default as waitFor } from './waitFor';
-export { default as waitForElementToBeRemoved } from './waitForElementToBeRemoved';
+export { default as waitFor } from './wait-for';
+export { default as waitForElementToBeRemoved } from './wait-for-element-to-be-removed';
 export { within, getQueriesForElement } from './within';
 
 export { configure, resetToDefaults } from './config';
@@ -12,7 +12,7 @@ export {
   isInaccessible,
 } from './helpers/accessiblity';
 export { getDefaultNormalizer } from './matches';
-export { renderHook } from './renderHook';
+export { renderHook } from './render-hook';
 export { screen } from './screen';
 export { userEvent } from './user-event';
 
@@ -22,6 +22,6 @@ export type {
   RenderResult as RenderAPI,
   DebugFunction,
 } from './render';
-export type { RenderHookOptions, RenderHookResult } from './renderHook';
+export type { RenderHookOptions, RenderHookResult } from './render-hook';
 export type { Config } from './config';
 export type { UserEventConfig } from './user-event';
