+++ src/queries/text.ts
@@ -1,5 +1,5 @@
 import type { ReactTestInstance } from 'react-test-renderer';
-import { findAll } from '../helpers/findAll';
+import { findAll } from '../helpers/find-all';
 import { isHostText } from '../helpers/host-component-names';
 import { matchTextContent } from '../helpers/matchers/match-text-content';
 import { TextMatch, TextMatchOptions } from '../matches';
