+++ src/render.tsx
@@ -5,10 +5,10 @@ import act from './act';
 import { addToCleanupQueue } from './cleanup';
 import { getConfig } from './config';
 import { getHostChildren } from './helpers/component-tree';
-import debugDeep, { DebugOptions } from './helpers/debugDeep';
-import debugShallow from './helpers/debugShallow';
+import debugDeep, { DebugOptions } from './helpers/debug-deep';
+import debugShallow from './helpers/debug-shallow';
 import { configureHostComponentNamesIfNeeded } from './helpers/host-component-names';
-import { validateStringsRenderedWithinText } from './helpers/stringValidation';
+import { validateStringsRenderedWithinText } from './helpers/string-validation';
 import { renderWithAct } from './render-act';
 import { setRenderResult, screen } from './screen';
 import { getQueriesForElement } from './within';
